+++ mediciones/forms.py
@@ -2,6 +2,8 @@ from django import forms
 from .choices import cumplimiento, tipo_medicion
 from .models import Medicion, InstrumentoMedicion
 from fiscalizacion.models import Fiscalizacion
+from services.utils.validators import validar_tamano_archivo
+
 
 
 class MedicionForm(forms.ModelForm):
@@ -45,14 +47,18 @@ class MedicionForm(forms.ModelForm):
 
     fiscalizacion = forms.ModelChoiceField(
         queryset=Fiscalizacion.objects.all(),
-        widget=forms.Select(attrs={'class': 'form-control'}),
-        label='Fiscalizacion'
+        widget=forms.HiddenInput(),
+        label='Fiscalización'
     )
 
+    def clean_foto(self):
+        foto = self.cleaned_data.get('foto')
+        return validar_tamano_archivo(foto, 100)
 
     class Meta:
         model = Medicion
         fields = '__all__'
+
