+++ proyectos/forms.py
@@ -1,7 +1,6 @@
 from django import forms
 from .choices import tipo_alumbrado
 from .models import Titular, RepresentanteLegal, DetalleLuminaria,  Proyecto
-from services.utils.validators import validar_tamano_archivo
 
 
 class ProyectoForm(forms.ModelForm):
@@ -36,11 +35,3 @@ class ProyectoForm(forms.ModelForm):
     class Meta:
         model = Proyecto
         fields = '__all__'
-
-    def clean_foto(self):
-        foto = self.cleaned_data.get('foto')
-        return validar_tamano_archivo(foto, 100)
-
-
-
-
