+++ templates/fiscalizacion/mantenedor_fiscalizacion.html
@@ -3,18 +3,7 @@
 
 {% block contenido %}
 <h1 class="text-center mb-5">Fiscalizaciones</h1>
-{{fiscalizaciones}}
-<tbody>
-    {% for fiscalizacion in fiscalizaciones %}
-    <tr>
-        <td>{{ fiscalizacion.id }}</td>
-        <td>{{ fiscalizacion.proyecto.nombre }}</td>
-        <td>{{ fiscalizacion.usuario.nombre }} {{ fiscalizacion.usuario.a_materno }}</td>
-        <td>{{ fiscalizacion.creado }}</td>
-        <!-- Agrega más columnas según sea necesario -->
-    </tr>
-    {% endfor %}
-</tbody>
+
 
 <div class="row">
 
@@ -35,9 +24,14 @@
         <div class="collapse" id="collapseExample">
             <div class="card card-body bg-dark mb-3">
                 <!-- Formulario para crear nueva fiscalización -->
-                <form method="post" action="{% url 'crear_fiscalizacion' %}">
+                <form method="post" enctype=multipart/form-data action="{% url 'crear_fiscalizacion' %}">
                     {% csrf_token %}
-                   
+                        <div class="col-3">
+                            <img src="{% static 'images/default.png' %}" alt="Imagen del área fiscalizada" width="150" height="auto"
+                                class="mx-auto my-auto d-block mb-2">
+                            <label for="foto">Fotografia del Proyecto:</label>
+                            <input type="file" name="foto" id="foto" class="form-control form-control-sm">
+                        </div>
                         <div class="row ps-5 pe-5">
                             <div class="mb-3 col-8">
                                 {{ form.proyecto.label_tag }}
@@ -93,28 +87,28 @@
                 </tr>
             </thead>
             <tbody>
-                 {% for fiscalizacion in fiscalizaciones %}
+                 {% for fiscalizacion in fiscalizaciones_con_mediciones %}
                 <tr>
-                    <td>{{ fiscalizacion.proyecto.nombre }}</td>
-                    <td>{{ fiscalizacion.id }}</td>
-                    <td>{{ fiscalizacion.proyecto.get_tipo_alumbrado_display }}</td>
-                    <td>a</td>
-                    <td>a</td>
-                    <td>a</td>
-                    <td>a</td>
+                    <td>{{ fiscalizacion.fiscalizacion.proyecto.nombre }}</td>
+                    <td>{{ fiscalizacion.fiscalizacion.id }}</td>
+                    <td>{{ fiscalizacion.fiscalizacion.proyecto.get_tipo_alumbrado_display }}</td>
+                    <td>{{ fiscalizacion.estadisticas.valor_minimo }}</td>
+                    <td>{{ fiscalizacion.estadisticas.valor_maximo }}</td>
+                    <td>{{ fiscalizacion.estadisticas.valor_promedio }}</td>
+                    <td>{{ fiscalizacion.mediciones.count }}</td>
                     <td>a{{ item.|default:"N/A"  }}</td>
                     <td>
                         <div class="d-flex justify-content-center">
                             <div class="" data-bs-toggle="tooltip" title="Ver detalle" data-bs-placement="top">
                                 <a class="btn btn-primary me-1"
-                                    href="{% url 'detalleFiscalizacion' fiscalizacion.id %}"><i
+                                    href="{% url 'detalleFiscalizacion' fiscalizacion.fiscalizacion.id %}"><i
                                         class="bi bi-eye"></i></a>
                             </div>
                             <div class="me-1" data-bs-toggle="tooltip" title="Nueva Medición" data-bs-placement="top">
-                                <a class="btn btn-success" href="{% url 'nuevaMedicion' fiscalizacion.id%}"><i class="bx bx-plus"></i></a>
+                                <a class="btn btn-success" href="{% url 'nuevaMedicion' fiscalizacion.fiscalizacion.id%}"><i class="bx bx-plus"></i></a>
                             </div>
                             <div data-bs-toggle="tooltip" title="Eliminar" data-bs-placement="top">
-                                <a class="btn btn-danger" href="{% url 'eliminarFiscalizacion' fiscalizacion.id %}"><i class="bx bx-trash"></i></a>
+                                <a class="btn btn-danger" href="{% url 'eliminarFiscalizacion' fiscalizacion.fiscalizacion.id %}"><i class="bx bx-trash"></i></a>
                             </div>
                         </div>
                     </td>
