+++ templates/usuarios/mantenedor_usuarios.html
@@ -2,7 +2,10 @@
 {% load static %}
 
 {% block contenido %}
-<h1 class="text-center mb-5">Usuarios</h1>
+<h1 class="text-center mb-5">Usuarios registrados</h1>
+
+
+
 
 <div class="d-flex justify-content-end">
